requirements.txt
numpy==1.25.2
Pillow==10.2.0
PyYAML==6.0.1
six==1.16.0
timm==0.6.7
torch==2.0.0+cu118
torchvision==0.15.2+cu118

